+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.6.0
 
-* Stream.bindCancellable: Added reportCancelledException parameter to emit CancelledException as an
+* Stream.bindCancellable: Added emitCancelledException parameter to emit CancelledException as an
   error in the stream.
 
 ## 2.5.0